Runtime: remove value_types
